+++ Realm/Realm/DatabaseTypes/RealmObjectBase.cs
@@ -20,16 +20,13 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 using System.Runtime.Serialization;
 using System.Xml.Serialization;
 using Realms.DataBinding;
-using Realms.Exceptions;
 using Realms.Helpers;
 using Realms.Native;
 using Realms.Schema;
@@ -37,328 +34,6 @@ using Realms.Weaving;
 
 namespace Realms
 {
-    internal static class RealmObjectExtensions
-    {
-        internal static RealmObjectBase.Metadata ObjectMetadata(this RealmObjectBase ro)
-        {
-            return ro.RealmAccessor().Metadata;
-        }
-
-        internal static ObjectHandle ObjectHandle(this RealmObjectBase ro)
-        {
-            return ro.RealmAccessor().ObjectHandle;
-        }
-
-        internal static ManagedAccessor RealmAccessor(this RealmObjectBase ro)
-        {
-            return ro.Accessor as ManagedAccessor;
-        }
-    }
-
-    internal class ManagedAccessor : IRealmAccessor, IThreadConfined
-    {
-        private bool _isEmbedded;
-
-        private Lazy<int> _hashCode;
-
-        private Realm _realm;
-
-        private ObjectHandle _objectHandle;
-
-        private RealmObjectBase.Metadata _metadata;
-
-        private NotificationTokenHandle _notificationToken;
-
-        internal ObjectHandle ObjectHandle => _objectHandle;
-
-        internal RealmObjectBase.Metadata ObjectMetadata => _metadata;
-
-        public bool IsManaged => true;
-
-        public bool IsValid => _objectHandle?.IsValid != false;
-
-        public bool IsFrozen => _realm?.IsFrozen == true;
-
-        public Realm Realm => _realm;
-
-        public ObjectSchema ObjectSchema => _metadata?.Schema;
-
-        public Lazy<int> HashCode => _hashCode;
-
-        public int BacklinksCount => _objectHandle?.GetBacklinkCount() ?? 0;
-
-        public IThreadConfinedHandle Handle => _objectHandle;
-
-        public RealmObjectBase.Metadata Metadata => _metadata;
-
-        [SuppressMessage("StyleCop.CSharp.NamingRules", "SA1300:Element should begin with upper-case letter", Justification = "This is the private event - the public is uppercased.")]
-        private event PropertyChangedEventHandler _propertyChanged;
-
-        public ManagedAccessor(Realm realm, ObjectHandle objectHandle, RealmObjectBase.Metadata metadata, bool isEmbedded)
-        {
-            _isEmbedded = isEmbedded;
-            _realm = realm;
-            _objectHandle = objectHandle;
-            _metadata = metadata;
-            _hashCode = new Lazy<int>(() => _objectHandle.GetObjHash());
-        }
-
-        public RealmObjectBase FreezeImpl()
-        {
-            var frozenRealm = Realm.Freeze();
-            var frozenHandle = ObjectHandle.Freeze(frozenRealm.SharedRealmHandle);
-            return frozenRealm.MakeObject(ObjectMetadata, frozenHandle);
-        }
-
-        public RealmValue GetValue(string propertyName)
-        {
-            return _objectHandle.GetValue(propertyName, _metadata, _realm);
-        }
-
-        public void SetValue(string propertyName, RealmValue val)
-        {
-            _objectHandle.SetValue(propertyName, _metadata, val, _realm);
-        }
-
-        public void SetValueUnique(string propertyName, RealmValue val)
-        {
-            if (_realm.IsInMigration)
-            {
-                _objectHandle.SetValue(propertyName, _metadata, val, _realm);
-            }
-            else
-            {
-                _objectHandle.SetValueUnique(propertyName, _metadata, val);
-            }
-        }
-
-        public IList<T> GetListValue<T>(string propertyName)
-        {
-            _metadata.Schema.TryFindProperty(propertyName, out var property);
-            return _objectHandle.GetList<T>(_realm, propertyName, _metadata, property.ObjectType);
-        }
-
-        public ISet<T> GetSetValue<T>(string propertyName)
-        {
-            _metadata.Schema.TryFindProperty(propertyName, out var property);
-            return _objectHandle.GetSet<T>(_realm, propertyName, _metadata, property.ObjectType);
-        }
-
-        public IDictionary<string, TValue> GetDictionaryValue<TValue>(string propertyName)
-        {
-            _metadata.Schema.TryFindProperty(propertyName, out var property);
-            return _objectHandle.GetDictionary<TValue>(_realm, propertyName, _metadata, property.ObjectType);
-        }
-
-        public IQueryable<T> GetBacklinks<T>(string propertyName)
-            where T : RealmObjectBase
-        {
-            var resultsHandle = _objectHandle.GetBacklinks(propertyName, _metadata);
-            return GetBacklinksForHandle<T>(propertyName, resultsHandle);
-        }
-
-        internal RealmResults<T> GetBacklinksForHandle<T>(string propertyName, ResultsHandle resultsHandle)
-            where T : RealmObjectBase
-        {
-            _metadata.Schema.TryFindProperty(propertyName, out var property);
-            var relatedMeta = _realm.Metadata[property.ObjectType];
-
-            return new RealmResults<T>(_realm, resultsHandle, relatedMeta);
-        }
-
-        public void SubscribeForNotifications()
-        {
-            Debug.Assert(_notificationToken == null, "_notificationToken must be null before subscribing.");
-
-            if (IsFrozen)
-            {
-                throw new RealmFrozenException("It is not possible to add a change listener to a frozen RealmObjectBase since it never changes.");
-            }
-
-            Realm.ExecuteOutsideTransaction(() =>
-            {
-                if (ObjectHandle.IsValid)
-                {
-                    var managedObjectHandle = GCHandle.Alloc(this, GCHandleType.Weak);
-                    _notificationToken = ObjectHandle.AddNotificationCallback(GCHandle.ToIntPtr(managedObjectHandle));
-                }
-            });
-        }
-
-        public void UnsubscribeFromNotifications()
-        {
-            _notificationToken?.Dispose();
-            _notificationToken = null;
-        }
-
-        public string GetStringDescription()
-        {
-            var typeString = GetType().Name;  //TODO THIS IS WRONG...
-
-            if (!IsManaged)  //TODO This can be removed
-            {
-                return $"{typeString} (unmanaged)";
-            }
-
-            if (!IsValid)
-            {
-                return $"{typeString} (removed)";
-            }
-
-            //if (!_isEmbedded && ObjectMetadata.Helper.TryGetPrimaryKeyValue(ro, out var pkValue))  //TODO We need to remove the helper from here
-            //{
-            //    var pkProperty = ObjectMetadata.Schema.PrimaryKeyProperty;
-            //    return $"{typeString} ({pkProperty.Value.Name} = {pkValue})";
-            //}
-
-            return typeString;
-        }
-    }
-
-    public interface IRealmAccessor
-    {
-        bool IsManaged { get; }
-
-        bool IsValid { get; }
-
-        bool IsFrozen { get; }
-
-        Realm Realm { get; }
-
-        ObjectSchema ObjectSchema { get; }
-
-        Lazy<int> HashCode { get; }  //TODO Should this be here? 
-
-        int BacklinksCount { get; }
-
-        public RealmObjectBase FreezeImpl();
-
-        RealmValue GetValue(string propertyName);
-
-        void SetValue(string propertyName, RealmValue val);
-
-        void SetValueUnique(string propertyName, RealmValue val);
-
-        IList<T> GetListValue<T>(string propertyName);
-
-        ISet<T> GetSetValue<T>(string propertyName);
-
-        IDictionary<string, TValue> GetDictionaryValue<TValue>(string propertyName);
-
-        IQueryable<T> GetBacklinks<T>(string propertyName) where T : RealmObjectBase;
-
-        void SubscribeForNotifications();
-
-        void UnsubscribeFromNotifications();
-
-        string GetStringDescription();
-    }
-
-    public class UnmanagedAccessor : IRealmAccessor
-    {
-        private Dictionary<string, object> _container = new();
-
-        public bool IsManaged => false;
-
-        public bool IsValid => true;
-
-        public bool IsFrozen => false;
-
-        public Realm Realm => null;
-
-        public ObjectSchema ObjectSchema => null;
-
-        public Lazy<int> HashCode => null;
-
-        public int BacklinksCount => 0;
-
-        public RealmObjectBase FreezeImpl()
-        {
-            throw new RealmException("Unmanaged objects cannot be frozen.");
-        }
-
-        public IQueryable<T> GetBacklinks<T>(string propertyName) where T : RealmObjectBase
-        {
-            Debug.Assert(false, "Object is not managed, but managed access was attempted");
-
-            throw new InvalidOperationException("Object is not managed, but managed access was attempted");
-        }
-
-        public IDictionary<string, TValue> GetDictionaryValue<TValue>(string propertyName)
-        {
-            if (!_container.ContainsKey(propertyName))
-            {
-                _container[propertyName] = new Dictionary<string, TValue>();
-            }
-
-            return (IDictionary<string, TValue>)_container[propertyName];
-        }
-
-        public IList<T> GetListValue<T>(string propertyName)
-        {
-            if (!_container.ContainsKey(propertyName))
-            {
-                _container[propertyName] = new List<T>();
-            }
-
-            return (IList<T>)_container[propertyName];
-        }
-
-        public ThreadSafeReference GetSafeReference()
-        {
-            Debug.Assert(false, "Object is not managed, but managed access was attempted");
-
-            throw new InvalidOperationException("Object is not managed, but managed access was attempted");
-        }
-
-        public ISet<T> GetSetValue<T>(string propertyName)
-        {
-            if (!_container.ContainsKey(propertyName))
-            {
-                _container[propertyName] = new HashSet<T>(RealmSet<T>.Comparer);
-            }
-
-            return (ISet<T>)_container[propertyName];
-        }
-
-        public string GetStringDescription()
-        {
-            Debug.Assert(false, "Object is not managed, but managed access was attempted");
-
-            throw new InvalidOperationException("Object is not managed, but managed access was attempted");
-        }
-
-        public RealmValue GetValue(string propertyName)
-        {
-            return (RealmValue)_container[propertyName];
-        }
-
-        public void SetValue(string propertyName, RealmValue val)
-        {
-            _container[propertyName] = val;
-        }
-
-        public void SetValueUnique(string propertyName, RealmValue val)
-        {
-            _container[propertyName] = val;
-        }
-
-        public void SubscribeForNotifications()
-        {
-            Debug.Assert(false, "Object is not managed, but managed access was attempted");
-        }
-
-        public void UnsubscribeFromNotifications()
-        {
-            Debug.Assert(false, "Object is not managed, but managed access was attempted");
-        }
-    }
-
-    public interface IRealmObject
-    {
-        IRealmAccessor Accessor { get; }
-    }
-
     /// <summary>
     /// Base for any object that can be persisted in a <see cref="Realm"/>.
     /// </summary>
