+++ Realm/Realm/DatabaseTypes/UnmanagedAccessor.cs
@@ -1,0 +1,127 @@
+﻿////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2022 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using Realms.Exceptions;
+using Realms.Schema;
+
+namespace Realms
+{
+    public class UnmanagedAccessor : IRealmAccessor
+    {
+        private Dictionary<string, object> _container = new();
+
+        public bool IsManaged => false;
+
+        public bool IsValid => true;
+
+        public bool IsFrozen => false;
+
+        public Realm Realm => null;
+
+        public ObjectSchema ObjectSchema => null;
+
+        public Lazy<int> HashCode => null;
+
+        public int BacklinksCount => 0;
+
+        public RealmObjectBase FreezeImpl()
+        {
+            throw new RealmException("Unmanaged objects cannot be frozen.");
+        }
+
+        public IQueryable<T> GetBacklinks<T>(string propertyName) where T : RealmObjectBase
+        {
+            Debug.Assert(false, "Object is not managed, but managed access was attempted");
+
+            throw new InvalidOperationException("Object is not managed, but managed access was attempted");
+        }
+
+        public IDictionary<string, TValue> GetDictionaryValue<TValue>(string propertyName)
+        {
+            if (!_container.ContainsKey(propertyName))
+            {
+                _container[propertyName] = new Dictionary<string, TValue>();
+            }
+
+            return (IDictionary<string, TValue>)_container[propertyName];
+        }
+
+        public IList<T> GetListValue<T>(string propertyName)
+        {
+            if (!_container.ContainsKey(propertyName))
+            {
+                _container[propertyName] = new List<T>();
+            }
+
+            return (IList<T>)_container[propertyName];
+        }
+
+        public ThreadSafeReference GetSafeReference()
+        {
+            Debug.Assert(false, "Object is not managed, but managed access was attempted");
+
+            throw new InvalidOperationException("Object is not managed, but managed access was attempted");
+        }
+
+        public ISet<T> GetSetValue<T>(string propertyName)
+        {
+            if (!_container.ContainsKey(propertyName))
+            {
+                _container[propertyName] = new HashSet<T>(RealmSet<T>.Comparer);
+            }
+
+            return (ISet<T>)_container[propertyName];
+        }
+
+        public string GetStringDescription()
+        {
+            Debug.Assert(false, "Object is not managed, but managed access was attempted");
+
+            throw new InvalidOperationException("Object is not managed, but managed access was attempted");
+        }
+
+        public RealmValue GetValue(string propertyName)
+        {
+            return (RealmValue)_container[propertyName];
+        }
+
+        public void SetValue(string propertyName, RealmValue val)
+        {
+            _container[propertyName] = val;
+        }
+
+        public void SetValueUnique(string propertyName, RealmValue val)
+        {
+            _container[propertyName] = val;
+        }
+
+        public void SubscribeForNotifications()
+        {
+            Debug.Assert(false, "Object is not managed, but managed access was attempted");
+        }
+
+        public void UnsubscribeFromNotifications()
+        {
+            Debug.Assert(false, "Object is not managed, but managed access was attempted");
+        }
+    }
+}
